LayoutTests/imported/w3c/web-platform-tests/content-security-policy/securitypolicyviolation/securitypolicyviolation-block-image.sub-expected.txt


FAIL Non-redirected same-origin URLs are not stripped. assert_equals: expected 4 but got 6
